"cspp1_m10_a1_1: 1/6 testcases passed. Cannot run hidden testcases as Server is unreachablepylint score: -1/10 "

+++ cspp1-assignments/CSPP1-Week3Exam/week3exam.py
@@ -56,10 +56,10 @@ def displaynew(adict, adict1):
 				# print(adictx)
 				for each1 in adictx:
 					total += int(adictx[each1])
-				score = (total/adict1[eachx])*100
+				score = int((total/adict1[eachx])*100)
 				# print(eachx)
 				# print(total)
-				print(score)
+				print(float(score))
 
 def main():
 		adict = {}
